Refactor watch button
